Rename passphrase > seed phrase to avoid confusion
