+++ app/src/main/java/com/concordium/wallet/ui/more/moreoverview/MoreOverviewFragment.kt
@@ -25,9 +25,9 @@ import com.concordium.wallet.ui.more.import.ImportActivity
 import com.concordium.wallet.ui.more.notifications.NotificationsPreferencesActivity
 import com.concordium.wallet.ui.more.tracking.TrackingPreferencesActivity
 import com.concordium.wallet.ui.more.unshielding.UnshieldingAccountsActivity
-import com.concordium.wallet.ui.passphrase.recoverprocess.RecoverProcessActivity
-import com.concordium.wallet.ui.passphrase.reveal.SavedPassPhraseRevealActivity
-import com.concordium.wallet.ui.passphrase.reveal.SavedSeedRevealActivity
+import com.concordium.wallet.ui.seed.recoverprocess.RecoverProcessActivity
+import com.concordium.wallet.ui.seed.reveal.SavedSeedPhraseRevealActivity
+import com.concordium.wallet.ui.seed.reveal.SavedSeedRevealActivity
 import com.concordium.wallet.ui.recipient.recipientlist.RecipientListActivity
 import com.concordium.wallet.ui.welcome.WelcomeActivity
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
@@ -297,7 +297,7 @@ class MoreOverviewFragment : BaseFragment() {
     }
 
     private fun revealSeedPhrase() {
-        val intent = Intent(activity, SavedPassPhraseRevealActivity::class.java)
+        val intent = Intent(activity, SavedSeedPhraseRevealActivity::class.java)
         startActivity(intent)
     }
