+++ app/src/main/java/com/concordium/wallet/ui/seed/recover/RecoverWalletActivity.kt
@@ -1,4 +1,4 @@
-package com.concordium.wallet.ui.passphrase.recover
+package com.concordium.wallet.ui.seed.recover
 
 import android.content.Intent
 import android.os.Bundle
@@ -10,17 +10,17 @@ import com.concordium.wallet.databinding.ActivityRecoverWalletBinding
 import com.concordium.wallet.ui.base.BaseActivity
 import com.concordium.wallet.ui.common.delegates.AuthDelegate
 import com.concordium.wallet.ui.common.delegates.AuthDelegateImpl
-import com.concordium.wallet.ui.passphrase.recoverprocess.RecoverProcessActivity
+import com.concordium.wallet.ui.seed.recoverprocess.RecoverProcessActivity
 import com.concordium.wallet.util.KeyboardUtil
 
 class RecoverWalletActivity :
-    BaseActivity(R.layout.activity_recover_wallet, R.string.pass_phrase_recover_title),
+    BaseActivity(R.layout.activity_recover_wallet, R.string.seed_phrase_recover_title),
     AuthDelegate by AuthDelegateImpl() {
 
     private val binding by lazy {
         ActivityRecoverWalletBinding.bind(findViewById(R.id.toastLayoutTopError))
     }
-    val viewModel: PassPhraseRecoverViewModel by lazy {
+    val viewModel: SeedPhraseRecoverViewModel by lazy {
         ViewModelProvider(
             this,
             ViewModelProvider.AndroidViewModelFactory.getInstance(application)
