+++ app/src/main/res/layout/activity_recover_wallet.xml
@@ -17,7 +17,7 @@
         android:layout_marginHorizontal="16dp"
         android:layout_marginTop="16dp"
         android:gravity="center"
-        android:text="@string/pass_phrase_recover_title"
+        android:text="@string/seed_phrase_recover_title"
         android:textAppearance="@style/CCX_Typography_H1"
         android:textColor="@color/ccx_neutral_tint_1" />
 
@@ -28,13 +28,13 @@
         android:layout_marginHorizontal="16dp"
         android:layout_marginTop="8dp"
         android:gravity="center"
-        android:text="@string/pass_phrase_recover_input_explain"
+        android:text="@string/seed_phrase_recover_input_explain"
         android:textAppearance="@style/CCX_Typography_Body"
         android:textColor="@color/ccx_neutral_tint_2" />
 
     <fragment
         android:id="@+id/input_fragment"
-        android:name="com.concordium.wallet.ui.passphrase.recover.PassPhraseRecoverInputFragment"
+        android:name="com.concordium.wallet.ui.seed.recover.SeedPhraseRecoverInputFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_marginTop="24dp" />
