+++ app/src/main/res/layout/activity_saved_seed_phrase_reveal.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ui.passphrase.setup.OneStepSetupWalletActivity">
+    tools:context=".ui.seed.setup.OneStepSetupWalletActivity">
 
     <include
         android:id="@+id/toolbar_layout"
@@ -39,7 +39,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:gravity="center"
-                android:text="@string/saved_pass_phrase_reveal_title"
+                android:text="@string/saved_seed_phrase_reveal_title"
                 android:textAppearance="@style/CCX_Typography_H1"
                 android:textColor="@color/ccx_neutral_tint_1"
                 app:layout_constraintTop_toTopOf="parent" />
@@ -49,7 +49,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginTop="8dp"
                 android:gravity="center"
-                android:text="@string/saved_pass_phrase_reveal_details"
+                android:text="@string/saved_seed_phrase_reveal_details"
                 android:textAppearance="@style/CCX_Typography_Body"
                 android:textColor="@color/ccx_neutral_tint_2" />
 
@@ -66,7 +66,7 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:drawablePadding="4dp"
-                    android:text="@string/saved_pass_phrase_reveal_transcribe"
+                    android:text="@string/saved_seed_phrase_reveal_transcribe"
                     android:textAppearance="@style/CCX_Typography_Caption"
                     android:textColor="@color/ccx_neutral_tint_1"
                     app:drawableStartCompat="@drawable/ccx_ico_approved" />
@@ -75,7 +75,7 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:drawablePadding="4dp"
-                    android:text="@string/saved_pass_phrase_reveal_digital_copy"
+                    android:text="@string/saved_seed_phrase_reveal_digital_copy"
                     android:textAppearance="@style/CCX_Typography_Caption"
                     android:textColor="@color/ccx_neutral_tint_1"
                     app:drawableStartCompat="@drawable/ccx_ico_declined" />
@@ -84,7 +84,7 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:drawablePadding="4dp"
-                    android:text="@string/saved_pass_phrase_reveal_screenshot"
+                    android:text="@string/saved_seed_phrase_reveal_screenshot"
                     android:textAppearance="@style/CCX_Typography_Caption"
                     android:textColor="@color/ccx_neutral_tint_1"
                     app:drawableStartCompat="@drawable/ccx_ico_declined" />
@@ -134,7 +134,7 @@
                 android:drawablePadding="8dp"
                 android:gravity="center"
                 android:paddingHorizontal="20dp"
-                android:text="@string/saved_pass_phrase_reveal_seed_phrase_copy"
+                android:text="@string/saved_seed_phrase_reveal_seed_phrase_copy"
                 android:textAppearance="@style/CCX_Typography_Button_M"
                 android:textColor="@color/ccx_neutral_tint_1"
                 app:drawableEndCompat="@drawable/cryptox_ico_copy"
@@ -148,7 +148,7 @@
                 android:drawablePadding="8dp"
                 android:gravity="center"
                 android:paddingHorizontal="20dp"
-                android:text="@string/saved_pass_phrase_reveal_seed_phrase_show"
+                android:text="@string/saved_seed_phrase_reveal_seed_phrase_show"
                 android:textAppearance="@style/CCX_Typography_Button_M"
                 android:textColor="@color/ccx_neutral_tint_1"
                 app:drawableEndCompat="@drawable/ccx_ico_eye_open"
