+++ app/src/main/res/layout/activity_saved_seed_reveal.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ui.passphrase.setup.OneStepSetupWalletActivity">
+    tools:context=".ui.seed.setup.OneStepSetupWalletActivity">
 
     <include
         android:id="@+id/toolbar_layout"
