+++ app/src/main/res/layout/fragment_recover_process_scanning.xml
@@ -10,7 +10,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:gravity="center"
-        android:text="@string/pass_phrase_recover_process_scanning"
+        android:text="@string/seed_phrase_recover_process_scanning"
         android:textAppearance="@style/CCX_Typography_Body"
         android:textColor="@color/ccx_neutral_tint_2"
         app:layout_constraintStart_toStartOf="parent"
@@ -21,7 +21,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="40dp"
-        android:text="@string/pass_phrase_recover_process_identities"
+        android:text="@string/seed_phrase_recover_process_identities"
         android:textAppearance="@style/CCX_Typography_Caption"
         android:textColor="@color/ccx_neutral_tint_1"
         app:layout_constraintEnd_toEndOf="parent"
@@ -44,7 +44,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="24dp"
-        android:text="@string/pass_phrase_recover_process_accounts"
+        android:text="@string/seed_phrase_recover_process_accounts"
         android:textAppearance="@style/CCX_Typography_Caption"
         android:textColor="@color/ccx_neutral_tint_1"
         app:layout_constraintEnd_toEndOf="parent"
