+++ app/src/main/res/layout/fragment_seed_phrase_recover_input.xml
@@ -13,7 +13,7 @@
         android:layout_marginEnd="4dp"
         android:gravity="start"
         android:padding="12dp"
-        android:text="@string/pass_phrase_recover_input_clear_all"
+        android:text="@string/seed_phrase_recover_input_clear_all"
         android:textAppearance="@style/CCX_Typography_Button_M"
         android:textColor="@color/ccx_neutral_tint_1"
         app:layout_constraintEnd_toStartOf="@id/btnClearBelow"
@@ -28,7 +28,7 @@
         android:layout_marginEnd="16dp"
         android:gravity="end"
         android:padding="12dp"
-        android:text="@string/pass_phrase_recover_input_clear_below"
+        android:text="@string/seed_phrase_recover_input_clear_below"
         android:textAppearance="@style/CCX_Typography_Button_M"
         android:textColor="@color/ccx_neutral_tint_1"
         app:layout_constraintEnd_toEndOf="parent"
@@ -138,7 +138,7 @@
         android:layout_marginEnd="16dp"
         android:drawablePadding="6sp"
         android:fontFamily="@font/w400"
-        android:text="@string/pass_phrase_try_another_word"
+        android:text="@string/seed_phrase_try_another_word"
         android:textColor="@color/cryptox_white_main"
         android:visibility="gone"
         app:layout_constraintBottom_toBottomOf="@id/words_list_view"
@@ -152,7 +152,7 @@
         style="@style/CryptoX_Container_Error.TextView"
         android:layout_marginHorizontal="16dp"
         android:layout_marginTop="24dp"
-        android:text="@string/pass_phrase_incorrect"
+        android:text="@string/seed_phrase_incorrect"
         android:textAppearance="@style/CCX_Typography_Body"
         android:visibility="gone"
         app:layout_constraintStart_toStartOf="parent"
@@ -175,7 +175,7 @@
             android:drawablePadding="6dp"
             android:gravity="center"
             android:padding="12dp"
-            android:text="@string/pass_phrase_recover_input_paste"
+            android:text="@string/seed_phrase_recover_input_paste"
             android:textAppearance="@style/CCX_Typography_Button_M"
             android:textColor="@color/ccx_neutral_tint_1"
             app:drawableStartCompat="@drawable/ccx_ico_notepad_16" />
