+++ app/src/main/res/values/strings.xml
@@ -563,33 +563,33 @@
 
     <string name="your_wallet_private_key">Your wallet private key</string>
     <string name="your_seed_phrase">Your seed phrase</string>
-    <string name="pass_phrase_incorrect">Incorrect secret recovery phrase. Please verify that each index has the right word.</string>
+    <string name="seed_phrase_incorrect">Incorrect secret recovery phrase. Please verify that each index has the right word.</string>
 
     <string name="copied">Copied</string>
 
-    <string name="pass_phrase_recover_title">@string/your_seed_phrase</string>
-    <string name="pass_phrase_recover_input_explain">Manually enter all your 24 words in order as it was presented to you</string>
-    <string name="pass_phrase_recover_input_clear_all">Clear all</string>
-    <string name="pass_phrase_recover_input_clear_below">Clear below</string>
-    <string name="pass_phrase_try_another_word">Please try another word</string>
-    <string name="pass_phrase_recover_input_paste">Paste your seed phrase</string>
+    <string name="seed_phrase_recover_title">@string/your_seed_phrase</string>
+    <string name="seed_phrase_recover_input_explain">Manually enter all your 24 words in order as it was presented to you</string>
+    <string name="seed_phrase_recover_input_clear_all">Clear all</string>
+    <string name="seed_phrase_recover_input_clear_below">Clear below</string>
+    <string name="seed_phrase_try_another_word">Please try another word</string>
+    <string name="seed_phrase_recover_input_paste">Paste your seed phrase</string>
 
-    <string name="pass_phrase_recover_process_title">Recover your accounts(s)</string>
-    <string name="pass_phrase_recover_process_subtitle">Recover your accounts and identities to connect them to your wallet</string>
-    <string name="pass_phrase_recover_process_scanning">Scanning the Concordium blockchain. Hang on while we find your accounts and identities.</string>
-    <string name="pass_phrase_recover_process_continue">Recover</string>
-    <string name="pass_phrase_recover_process_continue_to_wallet">Continue to wallet</string>
-    <string name="pass_phrase_recover_process_finished">Recovery finished</string>
-    <string name="pass_phrase_recover_process_partial">Recovery may be partial</string>
-    <string name="pass_phrase_recover_process_partial_description">There was a problem contacting some identity providers. This means that some identities and accounts may not have been recovered. If you can see that something is missing, you can try to restart the recovery process, or you can try again later by starting the recovery process from the settings menu. The affected identity providers:</string>
-    <string name="pass_phrase_recover_process_recovered">You have successfully recovered:</string>
-    <string name="pass_phrase_recover_process_nothing_new">Nothing new was found.</string>
-    <string name="pass_phrase_recover_process_try_again">Try again</string>
-    <string name="pass_phrase_recover_process_identities">Identities</string>
-    <string name="pass_phrase_recover_process_accounts">Accounts</string>
-    <string name="pass_phrase_recover_process_identity_name_prefix">Identity</string>
-    <string name="pass_phrase_recover_process_finished_no_accounts">No accounts</string>
-    <string name="pass_phrase_recover_process_finished_accounts_count">ACCOUNTS: %1$d</string>
+    <string name="seed_phrase_recover_process_title">Recover your accounts(s)</string>
+    <string name="seed_phrase_recover_process_subtitle">Recover your accounts and identities to connect them to your wallet</string>
+    <string name="seed_phrase_recover_process_scanning">Scanning the Concordium blockchain. Hang on while we find your accounts and identities.</string>
+    <string name="seed_phrase_recover_process_continue">Recover</string>
+    <string name="seed_phrase_recover_process_continue_to_wallet">Continue to wallet</string>
+    <string name="seed_phrase_recover_process_finished">Recovery finished</string>
+    <string name="seed_phrase_recover_process_partial">Recovery may be partial</string>
+    <string name="seed_phrase_recover_process_partial_description">There was a problem contacting some identity providers. This means that some identities and accounts may not have been recovered. If you can see that something is missing, you can try to restart the recovery process, or you can try again later by starting the recovery process from the settings menu. The affected identity providers:</string>
+    <string name="seed_phrase_recover_process_recovered">You have successfully recovered:</string>
+    <string name="seed_phrase_recover_process_nothing_new">Nothing new was found.</string>
+    <string name="seed_phrase_recover_process_try_again">Try again</string>
+    <string name="seed_phrase_recover_process_identities">Identities</string>
+    <string name="seed_phrase_recover_process_accounts">Accounts</string>
+    <string name="seed_phrase_recover_process_identity_name_prefix">Identity</string>
+    <string name="seed_phrase_recover_process_finished_no_accounts">No accounts</string>
+    <string name="seed_phrase_recover_process_finished_accounts_count">ACCOUNTS: %1$d</string>
 
     <string name="earn_title">Earning rewards</string>
     <string name="earn_subtitle">Below are two options for earning rewards on CryptoX Wallet:</string>
@@ -1008,14 +1008,14 @@
     <string name="setup_wallet_seed_phrase_screenshot">Screenshot</string>
     <string name="setup_wallet_seed_phrase_copied">@string/copied</string>
     <string name="setup_wallet_failed">Failed to set up the wallet, please contact \nsupport@concordium.software.com</string>
-    <string name="saved_pass_phrase_reveal_seed_phrase_copy">Copy to clipboard</string>
-    <string name="saved_pass_phrase_reveal_seed_phrase_show">Show seed phrase</string>
-    <string name="saved_pass_phrase_reveal_title">@string/your_seed_phrase</string>
-    <string name="saved_pass_phrase_reveal_details">Your seed phrase is the access key to all the funds in your wallet. If you forget it you will lose access to your wallet(s). Keep it somewhere safe</string>
-    <string name="saved_pass_phrase_reveal_transcribe">Transcribe</string>
-    <string name="saved_pass_phrase_reveal_digital_copy">Digital copy</string>
-    <string name="saved_pass_phrase_reveal_screenshot">Screenshot</string>
-    <string name="saved_pass_phrase_reveal_failed">Failed to reveal the phrase, please contact \nsupport@concordium.software.com</string>
+    <string name="saved_seed_phrase_reveal_seed_phrase_copy">Copy to clipboard</string>
+    <string name="saved_seed_phrase_reveal_seed_phrase_show">Show seed phrase</string>
+    <string name="saved_seed_phrase_reveal_title">@string/your_seed_phrase</string>
+    <string name="saved_seed_phrase_reveal_details">Your seed phrase is the access key to all the funds in your wallet. If you forget it you will lose access to your wallet(s). Keep it somewhere safe</string>
+    <string name="saved_seed_phrase_reveal_transcribe">Transcribe</string>
+    <string name="saved_seed_phrase_reveal_digital_copy">Digital copy</string>
+    <string name="saved_seed_phrase_reveal_screenshot">Screenshot</string>
+    <string name="saved_seed_phrase_reveal_failed">Failed to reveal the phrase, please contact \nsupport@concordium.software.com</string>
     <string name="saved_seed_reveal_seed_copy">Copy to clipboard</string>
     <string name="saved_seed_reveal_seed_show">Show wallet private key</string>
     <string name="saved_seed_reveal_title">@string/your_wallet_private_key</string>
